requirements.txt
yandex_tracker_client
openai
python-dotenv